requirements.txt
httpx
mcp
python-dotenv
Requests
starlette
urllib3
uvicorn
